+++ driver/src/main/java/oracle/nosql/driver/ops/PutResult.java
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2011, 2025 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2026 Oracle and/or its affiliates. All rights reserved.
  *
  * Licensed under the Universal Permissive License v 1.0 as shown at
  *  https://oss.oracle.com/licenses/upl/
@@ -110,8 +110,8 @@ public class PutResult extends WriteResult {
      *
      * @since 5.4.18
      */
-    public String getExistingRowMetadata() {
-        return super.getExistingRowMetadataInternal();
+    public String getExistingLastWriteMetadata() {
+        return super.getExistingLastWriteMetadataInternal();
     }
 
     /* from Result */