+++ driver/src/main/java/oracle/nosql/driver/ops/serde/nson/NsonSerializerFactory.java
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2011, 2025 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2026 Oracle and/or its affiliates. All rights reserved.
  *
  * Licensed under the Universal Permissive License v 1.0 as shown at
  *  https://oss.oracle.com/licenses/upl/
@@ -631,8 +631,8 @@ public class NsonSerializerFactory implements SerializerFactory {
             if (rq.getMatchVersion() != null) {
                 writeMapField(ns, ROW_VERSION, rq.getMatchVersion().getBytes());
             }
-            if (rq.getRowMetadata() != null) {
-                writeMapField(ns, ROW_METADATA, rq.getRowMetadata());
+            if (rq.getLastWriteMetadata() != null) {
+                writeMapField(ns, LAST_WRITE_METADATA, rq.getLastWriteMetadata());
             }
 
 
@@ -685,7 +685,7 @@ public class NsonSerializerFactory implements SerializerFactory {
             writeContinuationKey(ns, rq.getContinuationKey());
             writeFieldRange(ns, rq.getRange());
             writeKey(ns, rq);
-            writeMapField(ns, ROW_METADATA, rq.getRowMetadata());
+            writeMapField(ns, LAST_WRITE_METADATA, rq.getLastWriteMetadata());
             endMap(ns, PAYLOAD);
             ns.endMap(0); // top level object
         }
@@ -837,8 +837,8 @@ public class NsonSerializerFactory implements SerializerFactory {
             if (rq.getMatchVersion() != null) {
                 writeMapField(ns, ROW_VERSION, rq.getMatchVersion().getBytes());
             }
-            if (rq.getRowMetadata() != null) {
-                writeMapField(ns, ROW_METADATA, rq.getRowMetadata());
+            if (rq.getLastWriteMetadata() != null) {
+                writeMapField(ns, LAST_WRITE_METADATA, rq.getLastWriteMetadata());
             }
 
             /* writeValue uses the output stream directly */
@@ -1265,8 +1265,8 @@ public class NsonSerializerFactory implements SerializerFactory {
                 if (rq.getVirtualScan() != null) {
                     writeVirtualScan(ns, rq.getVirtualScan(), queryVersion);
                 }
-                if (rq.getRowMetadata() != null) {
-                    writeMapField(ns, ROW_METADATA, rq.getRowMetadata());
+                if (rq.getLastWriteMetadata() != null) {
+                    writeMapField(ns, LAST_WRITE_METADATA, rq.getLastWriteMetadata());
                 }
             }
 
@@ -3129,7 +3129,7 @@ public class NsonSerializerFactory implements SerializerFactory {
         }
 
         /**
-         * Reads the row from a get operation which includes row metadata
+         * Reads the row from a get operation which includes last write metadata
          * and the value
          */
         static void readRow(ByteInputStream in, GetResult result)
@@ -3150,8 +3150,8 @@ public class NsonSerializerFactory implements SerializerFactory {
                 } else if (name.equals(ROW_VERSION)) {
                     result.setVersion(Version.createVersion(
                                           Nson.readNsonBinary(in)));
-                } else if (name.equals(ROW_METADATA)) {
-                    result.setRowMetadata(Nson.readNsonString(in));
+                } else if (name.equals(LAST_WRITE_METADATA)) {
+                    result.setLastWriteMetadata(Nson.readNsonString(in));
                 } else if (name.equals(VALUE)) {
                     result.setValue((MapValue)Nson.readFieldValue(in));
                 } else {
@@ -3190,8 +3190,8 @@ public class NsonSerializerFactory implements SerializerFactory {
                     /* below requires change to WriteRequest */
                     // TODO } else if (name.equals(EXISTING_EXPIRATION)) {
                     //result.setExistingExpiration(Nson.readNsonLong(in));
-                } else if (name.equals(EXISTING_ROW_METADATA)) {
-                    result.setExistingRowMetadata(Nson.readNsonString(in));
+                } else if (name.equals(EXISTING_LAST_WRITE_METADATA)) {
+                    result.setExistingLastWriteMetadata(Nson.readNsonString(in));
                 } else {
                     skipUnknownField(walker, name);
                 }