+++ driver/src/test/java/oracle/nosql/driver/BasicTest.java
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2011, 2025 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2026 Oracle and/or its affiliates. All rights reserved.
  *
  * Licensed under the Universal Permissive License v 1.0 as shown at
  *  https://oss.oracle.com/licenses/upl/
@@ -1668,6 +1668,79 @@ public class BasicTest extends ProxyTestBase {
                 setTableName("flex");
             PutResult pres = handle.put(pr);
             assertNotNull(pres.getVersion());
+        }
+    }
+
+    @Test
+    public void testLastWriteMetadata() {
+
+        NoSQLHandle extraHandle = null;
+        try {
+            MapValue key = new MapValue().put("id", 10);
+            MapValue value = new MapValue().put("id", 10).put("name", "jane");
+
+            /* Create a table */
+            TableResult tres = tableOperation(
+                handle,
+                "create table if not exists testusers(id integer, " +
+                "name string, primary key(id))",
+                new TableLimits(500, 500, 50));
+            assertEquals(TableResult.State.ACTIVE, tres.getTableState());
+
+            final String jsonMetadata = "{\"abc\":\"def\"}";
+
+            /* PUT */
+            PutRequest putRequest = new PutRequest()
+                .setValue(value)
+                .setTableName("testusers")
+                .setLastWriteMetadata(jsonMetadata);
+
+            /*
+             * Note: this version check is just for pre-26.1.3 servers,
+             * Which do not support LWM. But after that, the support is
+             * based on the "features" flag in the proxy response header...
+             */
+            if (!checkKVVersion(26, 1, 3)) {
+                /* this should throw an error */
+                try {
+                    PutResult res = handle.put(putRequest);
+                    assertNotNull(res.getVersion());
+                } catch (Exception e) {
+                    /* success */
+                    return; /* no point in going forward */
+                }
+                fail("Last Write Metadata should have thrown error");
+            } else {
+                PutResult res = handle.put(putRequest);
+                assertNotNull(res.getVersion());
+            }
+
+            /* GET */
+            GetRequest getRequest = new GetRequest()
+                .setKey(key)
+                .setTableName("testusers");
+
+            GetResult res1 = handle.get(getRequest);
+            String lwm = res1.getLastWriteMetadata();
+            assertTrue(lwm != null && lwm.equals(jsonMetadata));
+
+            /* TODO: delete: how to verify LWM?? */
+
+            /*
+             * Create a new handle to verify that the "features" logic
+             * works correctly when the PutRequest is the very first
+             * request sent to the server
+             */
+            extraHandle = getHandle(endpoint);
+            PutResult res = extraHandle.put(putRequest);
+            assertNotNull(res.getVersion());
+        } catch (Exception e) {
+            e.printStackTrace();
+            fail("Exception in test");
+        } finally {
+            if (extraHandle != null) {
+                extraHandle.close();
+            }
         }
     }