+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [In progress]
 
 ### Changed
-- renamed get/set RowMetadata to get/set LastWriteMetadata.
+- Renamed get/set RowMetadata to get/set LastWriteMetadata.
+- Updated java minimum version to 11
+- Updated copyrights to 2026
 
 ## [5.4.19] 2025-12-23